refactor(mount): split mount command to many smaller ones for easier testing
